% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishtree-package.R
\docType{package}
\name{fishtree-package}
\alias{fishtree}
\alias{fishtree-package}
\title{fishtree: Interface to the Fish Tree of Life API}
\description{
fishtree makes it easy to access phylogenetic information about ray-finned
fishes with R. Download taxonomies, phylogenies, sequence matrices, and
diversification rate data with a simple set of functions.
}
\details{
Implementation note: this package makes calls over the network and caches
its (sometimes large) results for faster loading. Because of this,
long-running R instances could use a lot of memory.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fishtreeoflife.org/}
  \item Report bugs at \url{https://github.com/jonchang/fishtree/issues}
}

}
\author{
\strong{Maintainer}: Jonathan Chang \email{me@jonathanchang.org} (\href{https://orcid.org/0000-0002-3811-1254}{ORCID})

Other contributors:
\itemize{
  \item Michael E Alfaro [contributor]
  \item Daniel L Rabosky [contributor]
  \item Stephen A Smith [contributor]
}

}
\keyword{internal}
