% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fishtree_taxonomy}
\alias{fishtree_taxonomy}
\title{Download a taxonomy for the fish tree of life}
\usage{
fishtree_taxonomy(family = NULL, order = NULL)
}
\arguments{
\item{family}{retrieve one or more families}

\item{order}{retrieve one or more orders}
}
\value{
A list, with components containing data on the specified family or order.
}
\description{
Retrieves taxonomic and other information from the fish tree of life API. One of
\code{family} or \code{order} must be specified.
}
\examples{
test <- fishtree_taxonomy(family = "Labridae")
paste("There are ", length(test$sampled_species), "sampled species out of ", length(test$species), "in wrasses.")
}
