% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fishtree_alignment}
\alias{fishtree_alignment}
\title{Aligned sequences from the Fish Tree of Life}
\usage{
fishtree_alignment(name = NULL, split = FALSE)
}
\arguments{
\item{name}{Download an alignment for this rank (only families and orders are currently supported), or NULL for the alignment for all species.}

\item{split}{Split the alignment into a list of sequences by gene?}
}
\value{
An object of class \code{\link[ape]{DNAbin}}, or a named list of the same if \code{split = TRUE}
}
\description{
Retrieves an aligned sequence via the Fish Tree of Life API.
}
\references{
Rabosky, D. L., Chang, J., Title, P. O., Cowman, P. F., Sallan, L., Friedman, M., Kashner, K., Garilao, C., Near, T. J., Coll, M., Alfaro, M. E. (2018). An inverse latitudinal gradient in speciation rate for marine fishes. Nature, 559(7714), 392–395. doi:10.1038/s41586-018-0273-1
}
