% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fishtree_alignment}
\alias{fishtree_alignment}
\title{Download aligned sequences}
\usage{
fishtree_alignment(name = NULL, split = FALSE)
}
\arguments{
\item{name}{Download an alignment for this rank (only families and orders are currently supported), or NULL for the alignment for all species.}

\item{split}{Split the alignment into a list of sequences by gene?}
}
\value{
An object of class \code{DNAbin}
}
\description{
Retrieves an aligned sequence
}
\seealso{
[ape::DNAbin()]
}
