% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{fishtree_phylogeny}
\alias{fishtree_phylogeny}
\title{Download a phylogeny}
\usage{
fishtree_phylogeny(name = NULL, type = c("chronogram", "phylogram"))
}
\arguments{
\item{name}{Download phylogenies for this rank (only families and orders are currently supported)}

\item{type}{Either \code{chronogram} or \code{phylogram}}
}
\value{
An object of class \code{phylo}
}
\description{
Retrieves a phylogeny from the fish tree of life website.
}
\examples{
surgeons <- fishtree_phylogeny("Acanthuridae")
# Chronograms may not be ultrametric due to numerical precision issues
ape::is.ultrametric(surgeons)
ape::is.ultrametric(surgeons, tol = 0.00001)
}
\seealso{
[ape::read.tree()]
}
