% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fishtree_phylogeny}
\alias{fishtree_phylogeny}
\title{Phylogenies from the Fish Tree of Life}
\usage{
fishtree_phylogeny(name = NULL, type = c("chronogram", "phylogram"))
}
\arguments{
\item{name}{Download phylogenies for this rank (only families and orders are currently supported)}

\item{type}{Either \code{chronogram} or \code{phylogram}}
}
\value{
An object of class \code{phylo}
}
\description{
Retrieves a phylogeny via the Fish Tree of Life API
}
\examples{
surgeons <- fishtree_phylogeny("Acanthuridae")
# Chronograms may not be ultrametric due to numerical precision issues
ape::is.ultrametric(surgeons)
ape::is.ultrametric(surgeons, tol = 0.00001)
}
\references{
Rabosky, D. L., Chang, J., Title, P. O., Cowman, P. F., Sallan, L., Friedman, M., Kashner, K., Garilao, C., Near, T. J., Coll, M., Alfaro, M. E. (2018). An inverse latitudinal gradient in speciation rate for marine fishes. Nature, 559(7714), 392–395. doi:10.1038/s41586-018-0273-1
}
\seealso{
[ape::read.tree()]
}
