% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{fishtree_taxonomy}
\alias{fishtree_taxonomy}
\title{Download a taxonomy}
\usage{
fishtree_taxonomy(family = NULL, order = NULL)
}
\arguments{
\item{family}{retrieve one or more families}

\item{order}{retrieve one or more orders}
}
\value{
A list, with components containing data on the specified family or order.
}
\description{
Retrieves taxonomic information from the fish tree of life website. One of
\code{family} or \code{order} must be specified.
}
