% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fishtree_tip_rates}
\alias{fishtree_tip_rates}
\title{Retrieves tip rates for the fish tree of life}
\usage{
fishtree_tip_rates(name = NULL, sampled_only = TRUE)
}
\arguments{
\item{name}{optionally subset by a taxonomy rank (currently restricted to family and order)}

\item{sampled_only}{only include taxa actually present in the phylogeny?}
}
\value{
a data.frame
}
\description{
Downloads tip rates for the entire fish tree of life, or for a specified subset. Tip rates can be thought of as an
instantaneous speciation or extinction rate; for example, a higher tip-specific speciation rate might imply that
a lineage is more likely to split a new lineage at the present time.
}
\examples{
rates <- fishtree_tip_rates("Cichlidae")
plot(dr ~ lambda.tv, data = rates, main = "Do BAMM and DR speciation rates correlate?", xlab = "BAMM", ylab = "DR")
abline(0, 1, lty = 2, col = "red")
}
